+++ src/client/js/index.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:collection-timestampable/src/client/js/index.js
+ */
+
+import '../../common/js/index.js';